address comments: update tools
